Cleanup redundant cast warnings..
As a result of much of the warning cleanup work I've done over the past couple of years, the redundant cast warnings start happening more fequently.

   [repeat] /home/bwalker/src/netbeans/platform/core.multitabs/src/org/netbeans/core/multitabs/impl/ButtonPopupSwitcher.java:344: warning: [cast] redundant cast to Item
   [repeat]                 final Item item = ( Item ) pTable.getSelectedItem();
   [repeat]                                   ^

This works reduces those warnings.

+++ ide/css.editor/src/org/netbeans/modules/css/editor/module/spi/CssCompletionItem.java
@@ -649,7 +649,7 @@ public abstract class CssCompletionItem extends DefaultCompletionProposal {
                 if (mustBeEscaped) {
                     result.append("\\");
                     if (entry > 127 || Character.isWhitespace(entry) || Character.isISOControl(entry)) {
-                        result.append(String.format("%06x", (int) entry));
+                        result.append(String.format("%06x", entry));
                     } else {
                         result.appendCodePoint(entry);
                     }
